Clarify the supported versions of Windows.
It's made clear in the extended description that we only support versions of Windows that Microsoft supports, but it's useful to have that in the title too. Also - Window Server 2008 is still in support by Microsoft, even though its desktop equivalent (Windows Vista) is not. So this is actually a slight change to say we don't support Windows Server 2008 even though it's still in support by Microsoft.